// frontend/lib/api/generated/instructionResponseStatus.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Asya Backend API
 * API documentation for Asya Backend application
 * OpenAPI spec version: @appVersion@
 */

/**
 * 状态
 */
export type InstructionResponseStatus =
  (typeof InstructionResponseStatus)[keyof typeof InstructionResponseStatus];

export const InstructionResponseStatus = {
  SUBMITTED: "SUBMITTED",
  IN_PROGRESS: "IN_PROGRESS",
  REJECTED: "REJECTED",
  FEEDBACKED: "FEEDBACKED",
} as const;