// frontend/lib/api/generated/instructionReviewRequestStatus.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Asya Backend API
 * API documentation for Asya Backend application
 * OpenAPI spec version: @appVersion@
 */

/**
 * 目标状态: IN_PROGRESS, REJECTED, FEEDBACKED
 */
export type InstructionReviewRequestStatus =
  (typeof InstructionReviewRequestStatus)[keyof typeof InstructionReviewRequestStatus];

export const InstructionReviewRequestStatus = {
  SUBMITTED: "SUBMITTED",
  IN_PROGRESS: "IN_PROGRESS",
  REJECTED: "REJECTED",
  FEEDBACKED: "FEEDBACKED",
} as const;