// frontend/lib/api/generated/userGroupResponse.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Asya Backend API
 * API documentation for Asya Backend application
 * OpenAPI spec version: @appVersion@
 */

/**
 * 用户组响应
 */
export interface UserGroupResponse {
  /** 组ID */
  id: number;
  /** 组名称 */
  groupName: string;
  /** 成员UUID列表 */
  userUuids: string[];
}